+++ app/(lobby)/store/[id]/layout.tsx
