+++ app/(lobby)/store/[id]/loading.tsx
@@ -1,5 +1,3 @@
-import { AspectRatio } from "@/components/ui/aspect-ratio"
-import { Card, CardContent, CardFooter, CardHeader } from "@/components/ui/card"
 import { Skeleton } from "@/components/ui/skeleton"
 import { Icons } from "@/components/icons"
 import { Shell } from "@/components/shells/shell"
