+++ app/(lobby)/store/[id]/page.tsx
@@ -13,7 +13,8 @@ import { Shell } from "@/components/shells/shell"
 import ProductCard from "./components/product-card"
 import { SidebarNav } from "./components/sidebar-nav"
 
-const IndexPage = () => {
+const IndexPage = ({ params }: { params: { id: string } }) => {
+  console.log(params)
   const [isClient, setIsClient] = useState(false)
   const [activeCategory, setActiveCategory] = useState<number | null>(null)
   const sectionRefs = useRef<{ [key: string]: HTMLDivElement | null }>({})
@@ -22,7 +23,14 @@ const IndexPage = () => {
     setActiveCategory(index)
   }
 
-  const productListData = useSelector((state) => state.storeList.productList)
+  const storeData = useSelector((state) => {
+    return state.storeList.storeList.find((store) => store.id === params.id)
+  })
+
+  // 這裡使用了可選鏈 (?.) 和 Nullish Coalescing (??)
+  const productListData = useMemo(() => {
+    return storeData?.products ?? []
+  }, [storeData])
 
   const modifiedProductListData = useMemo(() => {
     const hotProducts = productListData.filter(
@@ -36,14 +44,15 @@ const IndexPage = () => {
     return [...hotProducts, ...otherProducts]
   }, [productListData])
 
+  // 使用 TypeScript 非空斷言（!）
   const productsByCategory = modifiedProductListData.reduce<{
     [key: string]: ProductType[]
   }>((acc, product) => {
     const { category } = product
-    if (!acc[category]) {
-      acc[category] = []
+    if (!acc[category!]) {
+      acc[category!] = []
     }
-    acc[category].push(product)
+    acc[category!].push(product)
     return acc
   }, {})
